integrated client and server communication

+++ lockfile-server/src/server.rs
@@ -8,16 +8,21 @@ use std::{
 };
 
 use lockfile_server::get_file_timestamp;
-use log::info;
 
 const SERVER_BUFFER_PATH: &str = "server/buffer";
 
 fn main() -> Result<(), Box<dyn Error>> {
-    // Create buffer in read-only mode
-    let server_buffer = OpenOptions::new()
+    env_logger::init();
+
+    // Create server buffer
+    let mut server_buffer = OpenOptions::new()
         .read(true)
+        .write(true)
         .create(true)
         .open(SERVER_BUFFER_PATH)?;
+
+    // Initialize server timestamp
+    let mut server_timestamp = get_file_timestamp(SERVER_BUFFER_PATH)?;
 
     // CtrlC handler. Removes lockfile
     ctrlc::set_handler(move || {
@@ -25,31 +30,46 @@ fn main() -> Result<(), Box<dyn Error>> {
         exit(130);
     })?;
 
-    // Watch server buffer for client connection
+    // Main loop
+    // Watches server buffer for client connection
     loop {
-        let server_timestamp = get_file_timestamp(SERVER_BUFFER_PATH)?;
-
         // Constant comparison of last file write timestamp
         if get_file_timestamp(SERVER_BUFFER_PATH)? != server_timestamp {
+            server_timestamp = get_file_timestamp(SERVER_BUFFER_PATH)?;
+
+            // Reopen the file
+            server_buffer = OpenOptions::new()
+                .read(true)
+                .write(true)
+                .create(true)
+                .open(SERVER_BUFFER_PATH)?;
+
             // Read lines from file with BufReader
             // First line is path to client file, used for response
             let reader = BufReader::new(&server_buffer);
+
             let lines = reader
                 .lines()
                 .map(|line| line.unwrap())
                 .collect::<Vec<String>>();
 
-            // Get client file path from first line
-            let client_file_path = &lines[0];
+            // Commit operation only if server buffer is not empty
+            if !lines.is_empty() {
+                // Get client file path from first line
+                let client_file_path = &lines[0];
 
-            // Construct response content from leftover lines
-            let response_content: String = lines[1..].to_vec().join("\n");
-            let _ = File::create(client_file_path)?.write(response_content.as_bytes())?;
+                // Construct response content from leftover lines
+                let response_content: String = lines[1..].to_vec().join("\n");
+                let _ = File::create(client_file_path)?.write(response_content.as_bytes())?;
 
-            info!("Written response to client");
+                println!("Written response to client");
 
-            sleep(Duration::from_millis(300));
-            let _ = fs::remove_file("server/lockfile");
+                // Truncate file for next client
+                server_buffer.set_len(0)?;
+
+                sleep(Duration::from_millis(300));
+                let _ = fs::remove_file("server/lockfile");
+            }
         }
     }
 }
